All changes for final project - separating word submissions into modules, setting up post routes, adding styles
