+++ app/index.html
@@ -1,26 +1,43 @@
+<!DOCTYPE html>
 <html>
+  <head>
+    <title>Band Name Generator</title>
+    <link href="style.css" rel="stylesheet" type="text/css">
+  </head>
   <body>
-    <h2>
+    <h1>
       Welcome to the Band Name Generator
-    </h2>
-
+    </h1>
     <button id="name">Get a random band name!</button>
-
-    <p>
-      Your name:
+    <p class="random-name">
       <span id="adjective"></span>
       <span id="verb"></span>
       <span id="noun"></span>
     </p>
 
     <br>
-    <h3>Submit words!</h3>
+    <h3>Submit your own words:</h3>
     <form id="submitWords">
-      <label>Give us an adjective:</label><input type="text" name="adjective">
-      <input type="submit" value="Submit Words"/>
+      <div>
+        <label>Give us an adjective:</label>
+        <input type="text" name="adjective">
+        <input type="submit" value="Submit Adjective"/>
+      </div>
+      <div>
+        <label>Give us a verb:</label>
+        <input type="text" name="verb">
+        <input type="submit" value="Submit Verb"/>
+      </div>
+      <div>
+        <label>Give us a noun:</label>
+        <input type="text" name="noun">
+        <input type="submit" value="Submit Noun"/>
+      </div>
     </form>
 
     <p id='adjectiveRes'></p>
+    <p id='verbRes'></p>
+    <p id='nounRes'></p>
 
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.2/jquery.min.js"></script>
     <script src="script.js"></script>
