+++ lib/noun.js
@@ -1,0 +1,16 @@
+'use strict';
+
+var Noun = function() {
+  this.ninjas = true;
+  this.cowboys = true;
+  this.bears = true;
+  this.beats = true;
+  this.alligators = true;
+  this.hamsters = true;
+  this.otters = true;
+  this.helicopters = true;
+  this.unicorns = true;
+  this.rhymes = true;
+};
+
+module.exports = Noun;
