+++ lib/postWord.js
@@ -1,0 +1,11 @@
+function postWord (word, wordObject) {
+  if (wordObject.hasOwnProperty(word)) {
+    return {msg: 'We already have your rad word, ' + word + ', in our list.'};
+  }
+
+  wordObject[word] = true;
+  console.dir(wordObject);
+  return {msg: 'Thanks for submitting ' + word + '!'};
+}
+
+module.exports = postWord;
